+++ src/components/graph/Axis.css
@@ -1,0 +1,4 @@
+.barChartAxis {
+  stroke: black;
+  stroke-width: 4px;
+}
