+++ src/components/graph/Bar.css
@@ -1,0 +1,3 @@
+.barChartRects {
+  fill: blue;
+}
