+++ src/components/graph/ConfidenceLines.css
@@ -1,0 +1,4 @@
+.confidenceLines {
+  stroke: red;
+  stroke-width: 4px;
+}
