+++ src/components/graph/XDivision.css
@@ -1,0 +1,3 @@
+.divisions {
+  stroke: grey;
+}
