+++ src/components/graph/XLabel.css
@@ -1,0 +1,5 @@
+.xLabel {
+  text-anchor: end;
+  alignment-baseline: middle;
+  fill: blue;
+}
