+++ src/components/graph/YLabel.css
@@ -1,0 +1,3 @@
+.yLabel {
+  transform: rotate(30deg);
+}
